+++ .gitignore
@@ -1,2 +1,2 @@
-/.project
+.project
 .DS_Store
