+++ app/views/settings/_code_theme_form.html.erb
@@ -1,4 +1,4 @@
-<fieldset class="box <%= 'hidden' unless HighlightJsSyntaxHighlighting.settings[:allow_redefine] %>">
+<fieldset class="box <%= 'hidden' unless RedmineHighlightjs.settings[:allow_redefine] %>">
 
 <legend><%=h l(:code_theme_title) %></legend>
 <div class="tabular">
