--- lib/code_theme_my_account_hooks.rb
@@ -1,3 +1,0 @@
-class CodeThemeMyAccountHooks < Redmine::Hook::ViewListener
-  render_on :view_my_account, partial: 'settings/code_theme_form', multipart: true
-end
